requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
streamlit>=1.40.0
pydantic>=2.8.0
pydantic-settings>=2.3.0
pandas>=2.2.0
rapidfuzz>=3.9.0
httpx>=0.27.0
requests>=2.32.0
python-dotenv>=1.0.1
openpyxl>=3.1.5
pytest>=8.3.0
transformers>=4.44.0
datasets>=2.20.0
peft>=0.12.0
accelerate>=0.33.0
torch>=2.3.0
scikit-learn>=1.5.0
sentencepiece>=0.2.0
bitsandbytes>=0.43.0; platform_system != "Windows"
